Refactor buffet_us and split files
